Break out Charge tests, and categorize StripeTest tests
